+++ config/shortcode-plus.php
@@ -16,4 +16,12 @@ return [
             ],
         ],
     ],
+
+    'livewire' => [
+        // Whether to use Livewire or not
+        'enabled' => true,
+
+        // Wether clicking on the image should open the modal or not
+        'modal' => true,
+    ],
 ];