+++ src/LaravelShortcodePlusServiceProvider.php
@@ -3,6 +3,8 @@
 namespace Murdercode\LaravelShortcodePlus;
 
 use Illuminate\Support\Facades\Blade;
+use Livewire\Livewire;
+use Murdercode\LaravelShortcodePlus\Http\Livewire\Modals\ModalImage;
 use Spatie\LaravelPackageTools\Package;
 use Spatie\LaravelPackageTools\PackageServiceProvider;
 
@@ -10,6 +12,11 @@ class LaravelShortcodePlusServiceProvider extends PackageServiceProvider
 {
     public function configurePackage(Package $package): void
     {
+        /*
+         * This class is a Package Service Provider
+         *
+         * More info: https://github.com/spatie/laravel-package-tools
+         */
         $package
             ->name('laravel-shortcode-plus')
             ->hasConfigFile()
@@ -25,6 +32,12 @@ class LaravelShortcodePlusServiceProvider extends PackageServiceProvider
 
     public function packageBooted()
     {
+        Blade::componentNamespace(
+            'Murdercode\LaravelShortcodePlus\View\Components',
+            'laravel-shortcode-plus'
+        );
+        Livewire::component('shortcode-plus::modals.modal-image', ModalImage::class);
+
         Blade::componentNamespace(
             'Murdercode\LaravelShortcodePlus\View\Components',
             'laravel-shortcode-plus'