+++ src/Parsers/Image.php
@@ -2,6 +2,7 @@
 
 namespace Murdercode\LaravelShortcodePlus\Parsers;
 
+use Murdercode\LaravelShortcodePlus\Helpers\ConfigHelper;
 use Murdercode\LaravelShortcodePlus\Helpers\ModelHelper;
 use Murdercode\LaravelShortcodePlus\Helpers\Sanitizer;
 
@@ -9,25 +10,29 @@ class Image
 {
     public static function parse(string $content): string
     {
+        $enable_modal = ConfigHelper::enableImageModal();
+
         return preg_replace_callback(
             '/\[image id="(\d+)"(?:\scaption="(.*?)")?(?:(.*?))\]/',
-            function ($matches) {
+            function ($matches) use ($enable_modal) {
                 $id_image = $matches[1];
                 $caption = $matches[2] ? Sanitizer::escapeQuotes($matches[2]) : null;
 
-                $image = ModelHelper::getInstance($id_image);
+                $model = new ModelHelper('image');
+                $image = $model->getModelClass()::find($id_image);
 
-                if (! $image) {
+                if (!$image) {
                     return 'Image not found';
                 }
 
-                $caption = $caption ?: ModelHelper::getValue($image, 'caption') ?: null;
-                $credits = ModelHelper::getValue($image, 'credits') ?: null;
-                $width = ModelHelper::getValue($image, 'width') ?: '1920';
-                $height = ModelHelper::getValue($image, 'height') ?: '1080';
-                $path = ModelHelper::getValue($image, 'path') ?: null;
-                $alternative_text = ModelHelper::getValue($image, 'alternative_text') ?: null;
-                $title = ModelHelper::getValue($image, 'title') ?: null;
+                $model->setModelInstance($image);
+                $caption = $caption ?: $model->getValueFromInstance('caption') ?: null;
+                $credits = $model->getValueFromInstance('credits') ?: null;
+                $width = $model->getValueFromInstance('width') ?: '1920';
+                $height = $model->getValueFromInstance('height') ?: '1080';
+                $path = $model->getValueFromInstance('path') ?: null;
+                $alternative_text = $model->getValueFromInstance('alternative_text') ?: null;
+                $title = $model->getValueFromInstance('title') ?: null;
 
                 return view(
                     'shortcode-plus::image',
@@ -38,7 +43,8 @@ class Image
                         'width',
                         'height',
                         'alternative_text',
-                        'title'
+                        'title',
+                        'enable_modal'
                     )
                 )->render();
             },