+++ tests/TestCase.php
@@ -3,6 +3,7 @@
 namespace Murdercode\LaravelShortcodePlus\Tests;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
+use Livewire\LivewireServiceProvider;
 use Murdercode\LaravelShortcodePlus\LaravelShortcodePlusServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
@@ -23,6 +24,7 @@ class TestCase extends Orchestra
     {
         return [
             LaravelShortcodePlusServiceProvider::class,
+            LivewireServiceProvider::class,
         ];
     }